requirements-linux.txt
cloudpickle==0.6.1
h5py==2.9.0
numpy==1.16.1
pathlib==1.0.1
protobuf==3.6.1
python-dateutil==2.7.5
python-utils==2.3.0
six==1.12.0
https://download.pytorch.org/whl/cpu/torch-1.0.0-cp36-cp36m-linux_x86_64.whl
torchfile==0.1.0
torchvision==0.2.1